pyproject.toml
[project]
name = "clawbot-plus"
version = "2.1.0"
description = "🤖 AI-Powered Browser & Computer Automation CLI with Voice, RAG, Multi-Agent & more"
authors = [{ name = "Aayushsoam" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
dependencies = [
    "browser-use>=0.12.0",
    "python-dotenv>=1.0.0",
    "InquirerPy>=0.3.4",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "pywin32>=306; sys_platform == 'win32'",
    "python-telegram-bot>=21.0",
    "ollama>=0.3.0",
    "pyperclip>=1.8.2",
    "edge-tts>=6.1.0",
    "SpeechRecognition>=3.10.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "plyer>=2.1.0",
]

[project.scripts]
clawbot = "clawbot.cli:main"

[project.urls]
Homepage = "https://github.com/aayushsoam/clawbot-plus"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clawbot", "skills"]
